facebook/api/requirements.txt
fastapi
uvicorn
python-dotenv
facebook-business
pydantic
gspread
google-auth
pandas